+++ sync.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"crypto/tls"
 	"encoding/binary"
+	"errors"
 	"io"
 	"net"
 	"sync"
@@ -28,7 +29,6 @@ import (
 	"github.com/loopholelabs/frisbee-go/internal/dialer"
 	"github.com/loopholelabs/frisbee-go/pkg/metadata"
 	"github.com/loopholelabs/frisbee-go/pkg/packet"
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 	"go.uber.org/atomic"
 )
@@ -141,7 +141,7 @@ func (c *Sync) RemoteAddr() net.Addr {
 //
 // If packet.Metadata.ContentLength == 0, then the content array must be nil. Otherwise, it is required that packet.Metadata.ContentLength == len(content).
 func (c *Sync) WritePacket(p *packet.Packet) error {
-	if int(p.Metadata.ContentLength) != len(*p.Content) {
+	if int(p.Metadata.ContentLength) != p.Content.Len() {
 		return InvalidContentLength
 	}
 
@@ -168,7 +168,7 @@ func (c *Sync) WritePacket(p *packet.Packet) error {
 		return c.closeWithError(err)
 	}
 	if p.Metadata.ContentLength != 0 {
-		_, err = c.conn.Write((*p.Content)[:p.Metadata.ContentLength])
+		_, err = c.conn.Write(p.Content.Bytes()[:p.Metadata.ContentLength])
 		if err != nil {
 			c.Unlock()
 			if c.closed.Load() {
@@ -208,11 +208,10 @@ func (c *Sync) ReadPacket() (*packet.Packet, error) {
 	p.Metadata.ContentLength = binary.BigEndian.Uint32(encodedPacket[metadata.ContentLengthOffset : metadata.ContentLengthOffset+metadata.ContentLengthSize])
 
 	if p.Metadata.ContentLength > 0 {
-		for cap(*p.Content) < int(p.Metadata.ContentLength) {
-			*p.Content = append((*p.Content)[:cap(*p.Content)], 0)
-		}
-		*p.Content = (*p.Content)[:p.Metadata.ContentLength]
-		_, err = io.ReadAtLeast(c.conn, *p.Content, int(p.Metadata.ContentLength))
+		contentLength := int(p.Metadata.ContentLength)
+		p.Content.Grow(contentLength)
+		p.Content.MoveOffset(contentLength)
+		_, err = io.ReadAtLeast(c.conn, p.Content.Bytes(), contentLength)
 		if err != nil {
 			if c.closed.Load() {
 				c.Logger().Debug().Err(ConnectionClosed).Msg("error while reading from underlying net.Conn")